Rename karma-mimic-reporter file
